+++ src/main/java/org/apiphany/security/oauth2/client/OAuth2ApiClient.java
@@ -2,7 +2,13 @@ package org.apiphany.security.oauth2.client;
 
 import static org.apiphany.ParameterFunction.parameter;
 
+import java.nio.charset.StandardCharsets;
+import java.security.PrivateKey;
+import java.security.Signature;
+import java.time.Instant;
+import java.util.Base64;
 import java.util.Map;
+import java.util.UUID;
 
 import org.apiphany.ApiClient;
 import org.apiphany.RequestParameters;
@@ -11,9 +17,12 @@ import org.apiphany.client.ExchangeClientBuilder;
 import org.apiphany.http.HttpAuthScheme;
 import org.apiphany.http.HttpHeader;
 import org.apiphany.io.ContentType;
+import org.apiphany.json.JsonBuilder;
+import org.apiphany.lang.Strings;
 import org.apiphany.security.AuthenticationException;
 import org.apiphany.security.AuthenticationToken;
 import org.apiphany.security.AuthenticationTokenProvider;
+import org.apiphany.security.MessageDigestAlgorithm;
 import org.apiphany.security.oauth2.AuthorizationGrantType;
 import org.apiphany.security.oauth2.ClientAuthenticationMethod;
 import org.apiphany.security.oauth2.OAuth2ClientRegistration;
@@ -30,6 +39,8 @@ import org.apiphany.security.oauth2.OAuth2ProviderDetails;
  */
 public class OAuth2ApiClient extends ApiClient implements AuthenticationTokenProvider {
 
+	private static final Base64.Encoder BASE64_URL_ENCODER = Base64.getUrlEncoder().withoutPadding();
+
 	/**
 	 * Configuration for the OAuth2 client registration.
 	 */
@@ -39,6 +50,16 @@ public class OAuth2ApiClient extends ApiClient implements AuthenticationTokenPro
 	 * Configuration details for the OAuth2 provider.
 	 */
 	private final OAuth2ProviderDetails providerDetails;
+
+	/**
+	 * The private key used when {@link ClientAuthenticationMethod#PRIVATE_KEY_JWT} is used.
+	 */
+	private PrivateKey privateKey;
+
+	/**
+	 * The signing algorithm used when {@link ClientAuthenticationMethod#PRIVATE_KEY_JWT} is used.
+	 */
+	private String signingAlgorithm;
 
 	/**
 	 * Constructs a new OAuth2 API client with the specified configurations.
@@ -69,6 +90,34 @@ public class OAuth2ApiClient extends ApiClient implements AuthenticationTokenPro
 	}
 
 	/**
+	 * Constructs a new OAuth2 API client with the specified configurations.
+	 *
+	 * @param clientRegistration the OAuth2 client registration details
+	 * @param providerDetails the OAuth2 provider configuration details
+	 * @param httpExchangeClient the HTTP exchange client to use for requests
+	 */
+	public OAuth2ApiClient(final OAuth2ClientRegistration clientRegistration, final OAuth2ProviderDetails providerDetails,
+			final PrivateKey privateKey, final String signingAlgorithm, final ExchangeClient httpExchangeClient) {
+		this(clientRegistration, providerDetails, httpExchangeClient);
+		this.privateKey = privateKey;
+		this.signingAlgorithm = signingAlgorithm;
+	}
+
+	/**
+	 * Constructs a new OAuth2 API client with the specified configurations.
+	 *
+	 * @param clientRegistration the OAuth2 client registration details
+	 * @param providerDetails the OAuth2 provider configuration details
+	 * @param exchangeClientBuilder the HTTP exchange client to use for requests
+	 */
+	public OAuth2ApiClient(final OAuth2ClientRegistration clientRegistration, final OAuth2ProviderDetails providerDetails,
+			final PrivateKey privateKey, final String signingAlgorithm, final ExchangeClientBuilder exchangeClientBuilder) {
+		this(clientRegistration, providerDetails, exchangeClientBuilder);
+		this.privateKey = privateKey;
+		this.signingAlgorithm = signingAlgorithm;
+	}
+
+	/**
 	 * Retrieves a standard authentication token.
 	 *
 	 * @return the authentication token, or null if the request fails
@@ -92,6 +141,8 @@ public class OAuth2ApiClient extends ApiClient implements AuthenticationTokenPro
 		return switch (method) {
 			case CLIENT_SECRET_BASIC -> getTokenWithClientSecretBasic();
 			case CLIENT_SECRET_POST -> getTokenWithClientSecretPost();
+			case CLIENT_SECRET_JWT -> getTokenWithClientSecretJwt();
+			case PRIVATE_KEY_JWT -> getTokenWithPrivateKeyJwt();
 			default -> throw new UnsupportedOperationException("Unsupported client authentication method: " + method);
 		};
 	}
@@ -105,6 +156,7 @@ public class OAuth2ApiClient extends ApiClient implements AuthenticationTokenPro
 		Map<String, String> params = RequestParameters.of(
 				parameter(OAuth2Parameter.GRANT_TYPE, clientRegistration.getAuthorizationGrantType()),
 				parameter(OAuth2Parameter.EXPIRES_IN, OAuth2Parameter.Default.EXPIRES_IN.toSeconds()));
+
 		return client()
 				.http()
 				.post()
@@ -127,6 +179,7 @@ public class OAuth2ApiClient extends ApiClient implements AuthenticationTokenPro
 				parameter(OAuth2Parameter.EXPIRES_IN, OAuth2Parameter.Default.EXPIRES_IN.toSeconds()),
 				parameter(OAuth2Parameter.CLIENT_ID, clientRegistration.getClientId()),
 				parameter(OAuth2Parameter.CLIENT_SECRET, clientRegistration.getClientSecret()));
+
 		return client()
 				.http()
 				.post()
@@ -135,5 +188,159 @@ public class OAuth2ApiClient extends ApiClient implements AuthenticationTokenPro
 				.header(HttpHeader.CONTENT_TYPE, ContentType.APPLICATION_FORM_URLENCODED)
 				.retrieve(AuthenticationToken.class)
 				.orRethrow(AuthenticationException::new);
+	}
+
+	/**
+	 * Returns the authentication token with {@link ClientAuthenticationMethod#CLIENT_SECRET_JWT}.
+	 *
+	 * @return the authentication token with `client_secret_jwt` method
+	 */
+	private AuthenticationToken getTokenWithClientSecretJwt() {
+		String clientAssertion = buildClientAssertionHmac(
+				clientRegistration.getClientId(),
+				providerDetails.getTokenUri(),
+				clientRegistration.getClientSecret());
+
+		Map<String, String> params = RequestParameters.of(
+				parameter(OAuth2Parameter.GRANT_TYPE, clientRegistration.getAuthorizationGrantType()),
+				parameter(OAuth2Parameter.CLIENT_ASSERTION_TYPE, "urn:ietf:params:oauth:client-assertion-type:jwt-bearer"),
+				parameter(OAuth2Parameter.CLIENT_ASSERTION, clientAssertion),
+				parameter(OAuth2Parameter.EXPIRES_IN, OAuth2Parameter.Default.EXPIRES_IN.toSeconds()));
+
+		return client()
+				.http()
+				.post()
+				.url(providerDetails.getTokenUri())
+				.body(RequestParameters.asString(RequestParameters.encode(params)))
+				.header(HttpHeader.CONTENT_TYPE, ContentType.APPLICATION_FORM_URLENCODED)
+				.retrieve(AuthenticationToken.class)
+				.orRethrow(AuthenticationException::new);
+	}
+
+	/**
+	 * Returns the authentication token with {@link ClientAuthenticationMethod#PRIVATE_KEY_JWT}.
+	 *
+	 * @return the authentication token with `private_key_jwt`.
+	 */
+	private AuthenticationToken getTokenWithPrivateKeyJwt() {
+		// Build a JWT signed with client private key (RSA or EC)
+		String clientAssertion = buildClientAssertionPrivateKey(
+				clientRegistration.getClientId(),
+				providerDetails.getTokenUri(),
+				privateKey,
+				signingAlgorithm);
+
+		Map<String, String> params = RequestParameters.of(
+				parameter(OAuth2Parameter.GRANT_TYPE, clientRegistration.getAuthorizationGrantType()),
+				parameter(OAuth2Parameter.CLIENT_ASSERTION_TYPE, "urn:ietf:params:oauth:client-assertion-type:jwt-bearer"),
+				parameter(OAuth2Parameter.CLIENT_ASSERTION, clientAssertion),
+				parameter(OAuth2Parameter.EXPIRES_IN, OAuth2Parameter.Default.EXPIRES_IN.toSeconds()));
+
+		return client()
+				.http()
+				.post()
+				.url(providerDetails.getTokenUri())
+				.body(RequestParameters.asString(RequestParameters.encode(params)))
+				.header(HttpHeader.CONTENT_TYPE, ContentType.APPLICATION_FORM_URLENCODED)
+				.retrieve(AuthenticationToken.class)
+				.orRethrow(AuthenticationException::new);
+	}
+
+	/**
+	 * Returns the client assertion HMAC.
+	 *
+	 * @param clientId the client ID
+	 * @param tokenEndpoint the token end point
+	 * @param clientSecret the client secret
+	 * @return the client assertion HMAC
+	 */
+	public static String buildClientAssertionHmac(final String clientId, final String tokenEndpoint, final String clientSecret) {
+		Map<String, Object> header = Map.of("alg", "HS256", "typ", "JWT");
+		Map<String, Object> claims = defaultClaims(clientId, tokenEndpoint);
+
+		String headerJson = Strings.removeAllWhitespace(JsonBuilder.toJson(header));
+		String payloadJson = Strings.removeAllWhitespace(JsonBuilder.toJson(claims));
+
+		String headerB64 = BASE64_URL_ENCODER.encodeToString(headerJson.getBytes(StandardCharsets.UTF_8));
+		String payloadB64 = BASE64_URL_ENCODER.encodeToString(payloadJson.getBytes(StandardCharsets.UTF_8));
+		String signingInput = headerB64 + "." + payloadB64;
+
+		byte[] signature = MessageDigestAlgorithm.SHA256.hmac(
+				clientSecret.getBytes(StandardCharsets.UTF_8), signingInput.getBytes(StandardCharsets.UTF_8));
+		return signingInput + "." + BASE64_URL_ENCODER.encodeToString(signature);
+	}
+
+	/**
+	 * Returns the client assertion private key.
+	 *
+	 * @param clientId the client ID
+	 * @param tokenEndpoint the token end point
+	 * @param privateKey the private key
+	 * @param algorithm the algorithm to sign with
+	 * @return the client assertion private key
+	 */
+	public static String buildClientAssertionPrivateKey(final String clientId, final String tokenEndpoint, final PrivateKey privateKey,
+			final String algorithm) {
+		// alg should be something like "RS256" or "ES256"
+		Map<String, Object> header = Map.of("alg", algorithm, "typ", "JWT");
+		Map<String, Object> claims = defaultClaims(clientId, tokenEndpoint);
+
+		String headerJson = Strings.removeAllWhitespace(JsonBuilder.toJson(header));
+		String payloadJson = Strings.removeAllWhitespace(JsonBuilder.toJson(claims));
+
+		String headerB64 = BASE64_URL_ENCODER.encodeToString(headerJson.getBytes(StandardCharsets.UTF_8));
+		String payloadB64 = BASE64_URL_ENCODER.encodeToString(payloadJson.getBytes(StandardCharsets.UTF_8));
+		String signingInput = headerB64 + "." + payloadB64;
+
+		byte[] signature = sign(privateKey, algorithm, signingInput.getBytes(StandardCharsets.UTF_8));
+		return signingInput + "." + BASE64_URL_ENCODER.encodeToString(signature);
+	}
+
+	/**
+	 * Returns the JWT default claims map.
+	 *
+	 * @param clientId the client ID
+	 * @param tokenEndpoint the token end point
+	 * @return the JWT default claims map
+	 */
+	private static Map<String, Object> defaultClaims(final String clientId, final String tokenEndpoint) {
+		Instant now = Instant.now();
+		return Map.of(
+				"iss", clientId,
+				"sub", clientId,
+				"aud", tokenEndpoint,
+				"jti", UUID.randomUUID().toString(),
+				"iat", now.getEpochSecond(),
+				"exp", now.plusSeconds(300).getEpochSecond() // 5 min
+		);
+	}
+
+	/**
+	 * Signs the input with the given private key and algorithm.
+	 *
+	 * @param privateKey the private key to sign with.
+	 * @param alg the algorithm used for signing
+	 * @param input the input bytes to sign.
+	 * @return signed output byte array
+	 */
+	private static byte[] sign(final PrivateKey privateKey, final String alg, final byte[] input) {
+		try {
+			String jcaAlg;
+			switch (alg) {
+				case "RS256" -> jcaAlg = "SHA256withRSA";
+				case "RS384" -> jcaAlg = "SHA384withRSA";
+				case "RS512" -> jcaAlg = "SHA512withRSA";
+				case "ES256" -> jcaAlg = "SHA256withECDSA";
+				case "ES384" -> jcaAlg = "SHA384withECDSA";
+				case "ES512" -> jcaAlg = "SHA512withECDSA";
+				default -> throw new IllegalArgumentException("Unsupported JWS alg: " + alg);
+			}
+			Signature sig = Signature.getInstance(jcaAlg);
+			sig.initSign(privateKey);
+			sig.update(input);
+			return sig.sign();
+		} catch (Exception e) {
+			throw new SecurityException("Error signing JWT", e);
+		}
 	}
 }